requirements.txt
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
locust==0.0
locustio==0.11.0
MarkupSafe==1.1.1
msgpack==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.8.0
pyzmq==18.0.1
requests==2.21.0
six==1.12.0
sortedcontainers==2.1.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.2
